a fucntion that sleeps with a loop and yield a random num

+++ python_async_comprehension/0-async_generator.py
@@ -3,14 +3,13 @@
 
 
 import asyncio
-from typing import Generator 
+from typing import Generator
 import random
 
 
-async def async_generator () -> Generator[int, None, None]:
+async def async_generator() -> Generator[float, None, None]:
     """a function that sleeps with loop"""
 
     for i in range(10):
         await asyncio.sleep(1)
-        yield random.uniform(0,10)
-
+        yield random.uniform(0, 10)
